+++ build.gradle.kts
@@ -1,6 +1,41 @@
-/*
- * This file was generated by the Gradle 'init' task.
- *
- * This is a general purpose Gradle build.
- * Learn more about Gradle by exploring our samples at https://docs.gradle.org/7.5/samples
- */
+plugins {
+    id("java")
+    id("net.minecrell.plugin-yml.bukkit") version "0.5.1"
+}
+
+group = "de.delta203"
+version = "1.0.0-SNAPSHOT"
+
+repositories {
+    mavenCentral()
+    maven("https://papermc.io/repo/repository/maven-public/")
+}
+
+dependencies {
+    compileOnly("io.papermc.paper:paper-api:1.19-R0.1-SNAPSHOT")
+}
+
+tasks {
+
+    compileJava {
+        options.release.set(17)
+        options.encoding = "UTF-8"
+    }
+
+    build {
+        dependsOn(jar)
+    }
+
+    jar {
+        archiveFileName.set("${rootProject.name}.${archiveExtension.getOrElse("jar")}")
+    }
+}
+
+bukkit {
+    main = "${rootProject.group}.chestsort.ChestSortPlugin"
+    apiVersion = "1.18"
+    name = "ChestSort"
+    load = net.minecrell.pluginyml.bukkit.BukkitPluginDescription.PluginLoadOrder.POSTWORLD
+    author = "Delta203"
+    authors = listOf("UniqueGame")
+}
