--- src/de/chestsort/delta203/plugin/SortType.java
@@ -1,6 +1,0 @@
-package de.chestsort.delta203.plugin;
-
-public enum SortType {
-
-	A_Z, AMOUNT,
-}
