+++ src/main/java/de/chestsort/delta203/plugin/SortType.java
@@ -1,0 +1,6 @@
+package de.chestsort.delta203.plugin;
+
+public enum SortType {
+
+	A_Z, AMOUNT,
+}
