--- src/plugin.yml
@@ -1,5 +1,0 @@
-name: ChestSort
-version: 1.0
-author: Delta203
-
-main: de.chestsort.delta203.plugin.ChestSort
